requirements.txt
gunicorn==21.2.0
pymysql
daphne==3.0.2
Django==4.2.7
django-celery-beat==2.1.0
#django-channels-presence==1.0.0
-e git+https://github.com/Cyparta/cyparta-django-channels-presence.git#egg=django-channels-presence
django-cleanup==5.1.0
django-cors-headers==3.4.0
django-mysql==3.12.0
django-rest-auth==0.9.5
django-safe-filefield==0.3.1
django-storages==1.11.1
django-timezone-field==4.0
djangorestframework==3.14.0
fcm-django==2.0.0
geoip2==4.4.0
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.1
Pillow==10.1.0
Random-Word==1.0.7
redis==3.5.3
SecretStorage==2.3.1
sqlparse==0.3.1
tweepy==4.5.0
Twisted==23.10.0
arrand==0.1
wget==3.2
whitenoise==6.6.0
psycopg2-binary==2.9.9
cryptography==41.0.5
channels-redis==3.4.1
boto3==1.28.75
opencv-python-headless==4.8.1.78
Unidecode==1.3.7
pyyaml
gevent==22.10.2
django-celery==3.1.17
django-import-export==2.7.1
django-soft-delete==0.9.21
django-simple-history==3.1.1
django-async-orm==0.1.10
phonenumbers==8.13.7
django-jsonfield==1.4.1
#collection==0.1.6
graphene==3.3
graphene-django==3.1.5
django-filter==23.4
reactivex
django_lifecycle
faker
CypartaGraphqlSubscriptionsTools
graphene-file-upload==1.3.0